[alpaka] Reorganise all code using ALPAKA_ACCELERATOR_NAMESPACE
Complete the reorganisation of the code using ALPAKA_ACCELERATOR_NAMESPACE.

All source and header files with code that defines or uses symbols in the
ALPAKA_ACCELERATOR_NAMESPACE should live in an .../alpkaka/ subdirectory:
  - files in an .../alpkaka/ subdirectory can include any other file;
  - files outside of an .../alpkaka/ subdirectory should not incliude
    header files from an .../alpkaka/ subdirectory.

Files with mixed symbols (i.e. files that define symbols both inside and
outside the ALPAKA_ACCELERATOR_NAMESPACE) have been split in two or more.
