+++ src/alpaka/AlpakaDataFormats/ZVertexHost.h
@@ -1,0 +1,10 @@
+#ifndef AlpakaDataFormats_ZVertexHost_h
+#define AlpakaDataFormats_ZVertexHost_h
+
+#include "AlpakaCore/alpakaConfig.h"
+#include "AlpakaCore/alpakaMemory.h"
+#include "AlpakaDataFormats/ZVertexSoA.h"
+
+using ZVertexHost = cms::alpakatools::host_buffer<ZVertexSoA>;
+
+#endif  // AlpakaDataFormats_ZVertexHost_h
