+++ src/alpaka/CondFormats/alpaka/SiPixelGainForHLTonGPU.h
@@ -1,5 +1,5 @@
-#ifndef CondFormats_SiPixelGainForHLTonGPU_h
-#define CondFormats_SiPixelGainForHLTonGPU_h
+#ifndef CondFormats_alpaka_SiPixelGainForHLTonGPU_h
+#define CondFormats_alpaka_SiPixelGainForHLTonGPU_h
 
 #include <utility>
 
@@ -59,4 +59,4 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
 
 }  // namespace ALPAKA_ACCELERATOR_NAMESPACE
 
-#endif  // CondFormats_SiPixelGainForHLTonGPU_h
+#endif  // CondFormats_alpaka_SiPixelGainForHLTonGPU_h
