+++ src/alpaka/plugin-BeamSpotProducer/alpaka/BeamSpotToAlpaka.cc
@@ -3,7 +3,7 @@
 #include "AlpakaCore/Product.h"
 #include "AlpakaCore/ScopedContext.h"
 #include "AlpakaCore/alpakaMemory.h"
-#include "AlpakaDataFormats/BeamSpotAlpaka.h"
+#include "AlpakaDataFormats/alpaka/BeamSpotAlpaka.h"
 #include "Framework/EDProducer.h"
 #include "Framework/Event.h"
 #include "Framework/EventSetup.h"
