+++ src/alpaka/plugin-PixelTrackFitting/alpaka/PixelTrackSoAFromAlpaka.cc
@@ -5,7 +5,8 @@
 #include "AlpakaCore/ScopedContext.h"
 #include "AlpakaCore/alpakaConfig.h"
 #include "AlpakaCore/alpakaMemory.h"
-#include "AlpakaDataFormats/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/alpaka/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/PixelTrackHost.h"
 #include "Framework/EDProducer.h"
 #include "Framework/Event.h"
 #include "Framework/EventSetup.h"
