+++ src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletAlpaka.cc
@@ -1,8 +1,8 @@
 #include "AlpakaCore/Product.h"
 #include "AlpakaCore/ScopedContext.h"
 #include "AlpakaCore/alpakaConfig.h"
-#include "AlpakaDataFormats/PixelTrackAlpaka.h"
-#include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
+#include "AlpakaDataFormats/alpaka/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/alpaka/TrackingRecHit2DAlpaka.h"
 #include "CAHitNtupletGeneratorOnGPU.h"
 #include "Framework/EDProducer.h"
 #include "Framework/Event.h"
