+++ src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletGeneratorKernels.h
@@ -4,7 +4,7 @@
 #include <algorithm>
 
 #include "AlpakaCore/alpakaMemory.h"
-#include "AlpakaDataFormats/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/alpaka/PixelTrackAlpaka.h"
 
 #include "GPUCACell.h"
