+++ src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletGeneratorOnGPU.cc
@@ -8,7 +8,7 @@
 #include <vector>
 
 #include "AlpakaCore/alpakaMemory.h"
-#include "AlpakaDataFormats/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/alpaka/PixelTrackAlpaka.h"
 #include "Framework/Event.h"
 
 #include "CAHitNtupletGeneratorOnGPU.h"
