+++ src/alpaka/plugin-PixelTriplets/alpaka/GPUCACell.h
@@ -14,8 +14,8 @@
 #include "AlpakaCore/VecArray.h"
 #include "AlpakaCore/alpakaConfig.h"
 #include "AlpakaCore/threadfence.h"
-#include "AlpakaDataFormats/PixelTrackAlpaka.h"
-#include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
+#include "AlpakaDataFormats/alpaka/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/alpaka/TrackingRecHit2DAlpaka.h"
 
 #include "../CAConstants.h"
 #include "../CircleEq.h"
