+++ src/alpaka/plugin-PixelTriplets/alpaka/RiemannFitOnGPU.h
@@ -9,7 +9,7 @@
 #include <cstdint>
 
 #include "AlpakaCore/alpakaConfig.h"
-#include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
+#include "AlpakaDataFormats/alpaka/TrackingRecHit2DAlpaka.h"
 #include "CondFormats/pixelCPEforGPU.h"
 
 #include "HelixFitOnGPU.h"
