+++ src/alpaka/plugin-SiPixelClusterizer/alpaka/SiPixelFedCablingMapESProducer.cc
@@ -7,7 +7,7 @@
 
 #include "AlpakaCore/alpakaConfig.h"
 #include "CondFormats/SiPixelFedCablingMapGPU.h"
-#include "CondFormats/SiPixelFedCablingMapGPUWrapper.h"
+#include "CondFormats/alpaka/SiPixelFedCablingMapGPUWrapper.h"
 #include "Framework/ESPluginFactory.h"
 #include "Framework/ESProducer.h"
 #include "Framework/EventSetup.h"
