+++ src/alpaka/plugin-SiPixelRecHits/alpaka/PixelRecHits.h
@@ -2,10 +2,10 @@
 #define plugin_SiPixelRecHits_alpaka_PixelRecHits_h
 
 #include "AlpakaCore/alpakaConfig.h"
-#include "AlpakaDataFormats/BeamSpotAlpaka.h"
-#include "AlpakaDataFormats/SiPixelClustersAlpaka.h"
-#include "AlpakaDataFormats/SiPixelDigisAlpaka.h"
-#include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
+#include "AlpakaDataFormats/alpaka/BeamSpotAlpaka.h"
+#include "AlpakaDataFormats/alpaka/SiPixelClustersAlpaka.h"
+#include "AlpakaDataFormats/alpaka/SiPixelDigisAlpaka.h"
+#include "AlpakaDataFormats/alpaka/TrackingRecHit2DAlpaka.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
