+++ src/alpaka/plugin-Validation/alpaka/HistoValidator.cc
@@ -10,11 +10,13 @@
 #include "AlpakaCore/ScopedContext.h"
 #include "AlpakaCore/alpakaConfig.h"
 #include "AlpakaCore/alpakaMemory.h"
-#include "AlpakaDataFormats/PixelTrackAlpaka.h"
-#include "AlpakaDataFormats/SiPixelClustersAlpaka.h"
-#include "AlpakaDataFormats/SiPixelDigisAlpaka.h"
-#include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
-#include "AlpakaDataFormats/ZVertexAlpaka.h"
+#include "AlpakaDataFormats/PixelTrackHost.h"
+#include "AlpakaDataFormats/ZVertexHost.h"
+#include "AlpakaDataFormats/alpaka/PixelTrackAlpaka.h"
+#include "AlpakaDataFormats/alpaka/SiPixelClustersAlpaka.h"
+#include "AlpakaDataFormats/alpaka/SiPixelDigisAlpaka.h"
+#include "AlpakaDataFormats/alpaka/TrackingRecHit2DAlpaka.h"
+#include "AlpakaDataFormats/alpaka/ZVertexAlpaka.h"
 #include "AlpakaDataFormats/gpuClusteringConstants.h"
 #include "Framework/EDProducer.h"
 #include "Framework/Event.h"
